youtube-summarizer/requirements.txt
google-generativeai>=0.3.0
youtube-transcript-api>=0.6.0
python-dotenv>=1.0.0
click>=8.0.0
requests>=2.25.0
urllib3>=1.26.0
colorama>=0.4.4
rich>=13.0.0
flask>=2.3.0
flask-cors>=4.0.0
flask-socketio>=5.3.0
gunicorn>=21.0.0
